+++ src/components/Header/index.tsx
@@ -2,9 +2,14 @@ import * as S from "./styled";
 import IconArrow from "@assets/icons/left-arrow.svg";
 import IconHome from "@assets/icons/home.svg";
 import { Link, useNavigate } from "react-router-dom";
-import PAGE_PATH from "../../../constants/path";
+import PAGE_PATH from "../../constants/path";
 
-const Header = () => {
+interface HeaderProps {
+  title: string;
+  hasHomeIcon?: boolean;
+}
+
+const Header = ({ title, hasHomeIcon = true }: HeaderProps) => {
   const navigate = useNavigate();
 
   const handleClickBackButton = () => {
@@ -16,10 +21,12 @@ const Header = () => {
       <S.NavigateBackButton onClick={handleClickBackButton}>
         <S.Icon src={IconArrow} alt="이전 페이지로 돌아가기" />
       </S.NavigateBackButton>
-      <S.Title>상황별 토픽</S.Title>
-      <Link to={PAGE_PATH.MAIN}>
-        <S.Icon src={IconHome} alt="메인 페이지로 이동하기" />
-      </Link>
+      <S.Title>{title}</S.Title>
+      {hasHomeIcon && (
+        <Link to={PAGE_PATH.MAIN}>
+          <S.Icon src={IconHome} alt="메인 페이지로 이동하기" />
+        </Link>
+      )}
     </S.Container>
   );
 };
