+++ src/components/Situation/index.tsx
@@ -1,11 +1,11 @@
 import { SITUATIONS } from "./data";
-import Header from "./Header";
+import Header from "../Header";
 import * as S from "./styled";
 
 const Situation = () => {
   return (
     <>
-      <Header />
+      <Header title="상황별 토픽" />
       <S.Main>
         <S.DescriptionBox>
           <S.Description>
